Split launch prompt context producers
